Prepare to publish as a rubygem
- Move everything into a lib/jsonerino directory
- Move entrypoint file to lib
- Add gemspec file
