+++ .gitignore
@@ -1,0 +1,4 @@
+Gemfile.lock
+pkg/*
+*.gem
+.bundle
