+++ lib/jsonerino/parse_error.rb
@@ -1,0 +1,3 @@
+module Jsonerino
+  class JsonParseError < StandardError; end
+end
