+++ lib/jsonerino/version.rb
@@ -1,0 +1,3 @@
+module Jsonerino
+  VERSION = '0.1.1'.freeze
+end
