--- lib/parse_error.rb
@@ -1,1 +1,0 @@
-class JsonParseError < StandardError; end
