--- main.rb
@@ -1,5 +1,0 @@
-#!/usr/bin/bash ruby
-
-require_relative './lib/lexer'
-
-puts 'JSON Parser/Linter'
